refactor: 폴더 오타 변경 exption -> exception
